feat: organize smart contracts file structures; moved Semaphore contracts to new structure
