+++ Cargo.toml
@@ -1,0 +1,10 @@
+[workspace]
+members = [
+    "contracts/group",
+    "contracts/semaphore_key_pts",
+    "contracts/verifier",
+    "contracts/semaphore"
+]
+
+[workspace.dependencies]
+soroban-sdk = "22.0.0-rc.3.2"
