+++ contracts/group/src/lib.rs
