+++ contracts/group/src/test.rs
