+++ contracts/semaphore/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "group"
+name = "hello-world"
 version = "0.0.0"
 edition = "2021"
 publish = false
