+++ contracts/semaphore_key_pts/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "semaphore_key_pts"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+doctest = false
+
+[dependencies]
+soroban-sdk = { workspace = true }
+
+[dev-dependencies]
+soroban-sdk = { workspace = true, features = ["testutils"] }
