+++ contracts/semaphore_key_pts/src/lib.rs
