+++ contracts/verifier/src/lib.rs
