--- semaphore-contracts/.gitignore
@@ -1,5 +1,0 @@
-# Rust's output directory
-target
-
-# Local Soroban settings
-.soroban
