--- semaphore-contracts/Cargo.toml
@@ -1,23 +1,0 @@
-[workspace]
-resolver = "2"
-members = [
-  "contracts/*",
-]
-
-[workspace.dependencies]
-soroban-sdk = "21.0.0"
-
-[profile.release]
-opt-level = "z"
-overflow-checks = true
-debug = 0
-strip = "symbols"
-debug-assertions = false
-panic = "abort"
-codegen-units = 1
-lto = true
-
-# For more information about this profile see https://soroban.stellar.org/docs/basic-tutorials/logging#cargotoml-profile
-[profile.release-with-logs]
-inherits = "release"
-debug-assertions = true
