--- semaphore-contracts/contracts/semaphore/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "hello-world"
-version = "0.0.0"
-edition = "2021"
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-doctest = false
-
-[dependencies]
-soroban-sdk = { workspace = true }
-
-[dev-dependencies]
-soroban-sdk = { workspace = true, features = ["testutils"] }
