--- semaphore-contracts/contracts/semaphore/src/test.rs
@@ -1,21 +1,0 @@
-#![cfg(test)]
-
-use super::*;
-use soroban_sdk::{vec, Env, String};
-
-#[test]
-fn test() {
-    let env = Env::default();
-    let contract_id = env.register_contract(None, HelloContract);
-    let client = HelloContractClient::new(&env, &contract_id);
-
-    let words = client.hello(&String::from_str(&env, "Dev"));
-    assert_eq!(
-        words,
-        vec![
-            &env,
-            String::from_str(&env, "Hello"),
-            String::from_str(&env, "Dev"),
-        ]
-    );
-}
